Migrate to webpack 5
DISCLAIMER : this migration is tested on promomaker (pellov). The
changes related to the polyfills of node modules may not be enough. It
depends on the packages used by the project.

During migration from 4 to 5, it is recommended that all plugins,
loaders, babel plugins, etc are up to date.

The major changes are

1. html-webpack-plugin

Our custom plugin HtmlPlugin is not working anymore(at least not solid
enough). It is replaced by the well used html-webpack-plugin.

Our renderHtml is re-used to provide the plugin a templateContent.

html-webpack-harddisk-plugin is used alongside to always
output the html file to disk which can then be served by the
devServer (via contentBase).

If the html file is not written to disk, it will be considered by the
devServer as an asset (like the js bundle) and will be served at the
publicpath (/assets/) like all the other assets. So the path in the
scripts and links won't be working (404).

2. vendor caching

To separate vendor chunk from our main bundle:
https://webpack.js.org/guides/caching/#extracting-boilerplate

3. polyfills of node modules

Webpack 5 do not provide anymore polyfills for node modules. They
recommend the use of browser-compatible packages or librairies.

We must provide ourselves the necessary polyfills by using this method:
https://webpack.js.org/guides/shimming/#granular-shimming

In our case, in react-markdown, a package called vfile needs to access
the process object. So it can be polyfilled with the package process
which provide a browser process.

Related issue: https://github.com/vfile/vfile/issues/38#issuecomment-683198538

Same issue with path, polyfilled with path-browserify. This fix (with
the resolve.fallback) was directly logged by webpack to the console.

4. webpack cli and webpack-dev-server

- to run the devServer, the command is now "webpack serve"
- for webpack-cli the --info-verbosity parameter is deprecated